// sdks/node/model/auth2faEnablePost200Response.ts
/**
 * HookSniff API
 * Webhook delivery, monitoring, and management API. All endpoints under `/v1` require authentication via `Authorization: Bearer <api_key>` header unless marked as **Public**. 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@hooksniff.vercel.app
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class Auth2faEnablePost200Response {
    'secret'?: string;
    'qrUrl'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "secret",
            "baseName": "secret",
            "type": "string"
        },
        {
            "name": "qrUrl",
            "baseName": "qr_url",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return Auth2faEnablePost200Response.attributeTypeMap;
    }
}